move log fns to utils/log/
