+++ admin/tabs/userData/roles/graph/RoleGraph.ts
@@ -14,7 +14,8 @@ import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {toolbar} from '@xh/hoist/desktop/cmp/toolbar';
 import {Icon} from '@xh/hoist/icon';
 import {Highcharts} from '@xh/hoist/kit/highcharts';
-import {logError, pluralize} from '@xh/hoist/utils/js';
+import {pluralize} from '@xh/hoist/utils/js';
+import {logError} from '@xh/hoist/utils/log';
 import {isEmpty} from 'lodash';
 import {RoleModel} from '../RoleModel';
 import {RoleGraphModel} from './RoleGraphModel';
