+++ core/HoistComponent.ts
@@ -26,7 +26,7 @@ import {
     HoistModel
 } from './model';
 import {throwIf, warnIf, withDefault} from '@xh/hoist/utils/js';
-import {logError} from '@xh/hoist/utils/js/LogUtils';
+import {logError} from '@xh/hoist/utils/log';
 import {getLayoutProps, useOnMount, useOnUnmount} from '@xh/hoist/utils/react';
 import classNames from 'classnames';
 import {isFunction, isPlainObject, isObject} from 'lodash';
