+++ desktop/cmp/button/zoneGrid/ZoneMapperButton.ts
@@ -12,7 +12,8 @@ import {ZoneMapperModel} from '@xh/hoist/cmp/zoneGrid/impl/ZoneMapperModel';
 import {zoneMapper} from '@xh/hoist/desktop/cmp/zoneGrid/impl/ZoneMapper';
 import {Icon} from '@xh/hoist/icon';
 import {popover, Position} from '@xh/hoist/kit/blueprint';
-import {logError, stopPropagation, withDefault} from '@xh/hoist/utils/js';
+import {logError} from '@xh/hoist/utils/log';
+import {stopPropagation, withDefault} from '@xh/hoist/utils/js';
 import {button, ButtonProps} from '../Button';
 
 export interface ZoneMapperButtonProps extends ButtonProps {
