+++ desktop/hooks/UseContextMenu.ts
@@ -7,7 +7,7 @@
 import type {ContextMenuSpec} from '@xh/hoist/core';
 import {contextMenu} from '@xh/hoist/desktop/cmp/contextmenu/ContextMenu';
 import {showContextMenu} from '@xh/hoist/kit/blueprint';
-import {logError} from '@xh/hoist/utils/js';
+import {logError} from '@xh/hoist/utils/log';
 import {isArray, isEmpty, isFunction, isUndefined} from 'lodash';
 import {cloneElement, isValidElement, MouseEvent, ReactElement} from 'react';
