+++ mobile/cmp/button/grid/ColChooserButton.ts
@@ -9,7 +9,8 @@ import {hoistCmp, useContextModel} from '@xh/hoist/core';
 import {Icon} from '@xh/hoist/icon';
 import {button, ButtonProps} from '@xh/hoist/mobile/cmp/button';
 import '@xh/hoist/mobile/register';
-import {logError, withDefault} from '@xh/hoist/utils/js';
+import {logError} from '@xh/hoist/utils/log';
+import {withDefault} from '@xh/hoist/utils/js';
 
 export interface ColChooserButtonProps extends ButtonProps {
     /** GridModel to which this button should bind. Will find nearest in context if not provided. */
